[refactor] 응답 관련 클래스들 코드 리팩토링 #9
