+++ src/main/java/com/triptune/global/response/ApiResponse.java
@@ -1,5 +1,6 @@
 package com.triptune.global.response;
 
+import com.triptune.global.enumclass.SuccessCode;
 import lombok.Getter;
 import lombok.Setter;
 import org.springframework.http.HttpStatus;
@@ -9,50 +10,16 @@ import org.springframework.util.MultiValueMap;
 @Getter
 @Setter
 public class ApiResponse<T> extends ResponseEntity<SuccessResponse<T>> {
-
     public static <T> ApiResponse<T> okResponse(){
-        SuccessResponse<T> successResponse = SuccessResponse.<T>builder()
-                .message("200(성공)")
-                .build();
-
-        return new ApiResponse<>(successResponse, HttpStatus.OK);
-    }
-
-    public static <T> ApiResponse<T> okResponse(String message){
-        SuccessResponse<T> successResponse = SuccessResponse.<T>builder()
-                .message(message)
-                .build();
-
-        return new ApiResponse<>(successResponse, HttpStatus.OK);
+        return new ApiResponse<>(SuccessResponse.of(), SuccessCode.GENERAL_SUCCESS.getStatus());
     }
 
     public static <T> ApiResponse<T> dataResponse(T body){
-        SuccessResponse<T> successResponse = SuccessResponse.<T>builder()
-                .message("200(성공)")
-                .data(body)
-                .build();
-
-        return new ApiResponse<>(successResponse, HttpStatus.OK);
-    }
-
-
-    public ApiResponse(HttpStatus status) {
-        super(status);
+        return new ApiResponse<>(SuccessResponse.of(body), SuccessCode.GENERAL_SUCCESS.getStatus());
     }
 
     public ApiResponse(SuccessResponse<T> body, HttpStatus status) {
         super(body, status);
     }
 
-    public ApiResponse(MultiValueMap<String, String> headers, HttpStatus status) {
-        super(headers, status);
-    }
-
-    public ApiResponse(SuccessResponse<T> body, MultiValueMap<String, String> headers, HttpStatus status) {
-        super(body, headers, status);
-    }
-
-    public ApiResponse(SuccessResponse<T> body, MultiValueMap<String, String> headers, int rawStatus) {
-        super(body, headers, rawStatus);
-    }
 }
